moving this 3 actors from fvtk to actor
